move macro to its own target as suggested by fabian
