+++ Tests/OracleNIOMacrosTests/OracleStatementMacroTests.swift
@@ -21,7 +21,7 @@ import XCTest
 
 // Macro implementations build for the host, so the corresponding module is not available when cross-compiling. Cross-compiled tests may still make use of the macro itself in end-to-end tests.
 #if canImport(OracleNIOMacros)
-    import OracleNIOMacros
+    import OracleNIOMacrosPlugin
 
     let testMacros: [String: MacroSpec] = [
         "Statement": MacroSpec(
